Splits test files per class.
