+++ aimmo/tests/tests_views.py
@@ -1,12 +1,11 @@
-import logging
 import ast
 import json
 
-from django.contrib.auth.models import AnonymousUser, User
+from django.contrib.auth.models import User
 from django.core.urlresolvers import reverse
 from django.test import Client, TestCase
 
-from aimmo import models, forms, app_settings
+from aimmo import models, app_settings
 
 app_settings.GAME_SERVER_URL_FUNCTION = lambda game_id: ('base %s' % game_id, 'path %s' % game_id)
 app_settings.GAME_SERVER_PORT_FUNCTION = lambda game_id: 0
@@ -255,66 +254,3 @@ class TestViews(TestCase):
         self.assertEqual(current_avatar_id, 1)
         self.assertEqual(len(games_api_users), 1)
         self.assertEqual(games_api_users[0]['id'], 1)
-
-
-class TestModels(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.user1 = User.objects.create_user('test', 'test@example.com', 'password')
-        cls.user1.save()
-        cls.user2 = User.objects.create_user('test2', 'test2@example.com', 'password')
-        cls.user2.save()
-        cls.game = models.Game(id=1, name='test', public=False)
-        cls.game.save()
-
-    def setUp(self):
-        self.game.refresh_from_db()
-
-    def test_public_games_can_be_accessed(self):
-        self.game.public = True
-        self.assertTrue(self.game.can_user_play(self.user1))
-        self.assertTrue(self.game.can_user_play(self.user2))
-
-    def test_anon_user_can_play_public_game(self):
-        self.game.public = True
-        self.assertTrue(self.game.can_user_play(AnonymousUser()))
-
-    def test_authed_user_can_play(self):
-        self.game.public = False
-        self.game.can_play = [self.user1]
-        self.assertTrue(self.game.can_user_play(self.user1))
-
-    def test_non_authed_user_cannot_play(self):
-        self.game.public = False
-        self.game.can_play = [self.user1]
-        self.assertFalse(self.game.can_user_play(self.user2))
-
-    def test_game_active_by_default(self):
-        self.assertTrue(self.game.is_active)
-
-    def test_completed_game_inactive(self):
-        self.game.completed = True
-        self.game.save()
-        self.assertFalse(self.game.is_active)
-
-
-class TestForms(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.user1 = User.objects.create_user('test', 'test@example.com', 'password')
-        cls.user1.save()
-        cls.game = models.Game(id=1, name='test', public=False)
-        cls.game.save()
-
-    def test_create_game(self):
-        form = forms.AddGameForm({'name': 'test2'})
-        form.add_playable_games({self.game})
-        self.assertTrue(form.is_valid())
-        new_game = form.save()
-        self.assertEqual(new_game.name, "test2")
-
-    def test_cannot_create_duplicate_game(self):
-        form = forms.AddGameForm({'name': 'test'})
-        form.add_playable_games({self.game})
-        self.assertFalse(form.is_valid())
-        self.assertEqual(form.errors, {'__all__': [u'Sorry, a game with this name already exists.']})
